requirements.txt
altgraph==0.16.1
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
fake-useragent==0.1.11
future==0.17.1
idna==2.8
lxml==4.4.1
pefile==2019.4.18
Pillow==6.1.0
PyInstaller==3.5
pyperclip==1.7.0
pywin32-ctypes==0.2.0
requests==2.22.0
soupsieve==1.9.3
urllib3==1.25.5
